Refactor project structure and update build configurations
Removed the outdated 'main.cxx' file and renamed 'main_1.cpp' to be more descriptive. Adjusted 'CMakeLists.txt' to use the updated C++23 standard and switched to a project structure. Updated 'CMakePresets.json' to include more building presets and set the C++ standard to 23.

The change in project structure is to consolidate source files under a single directory, making the project easier to navigate. The build configurations update is done to utilize the newer functionalities of C++23, and providing more options for building. It will enhance code organization and compatibility.

Signed-off-by: Evgeniy Gleba <60469435+geugenm@users.noreply.github.com>